DeviceTree: Get spintable entry from device tree
The applications processors will spin on a predefined entry in the so
called spintable as long as its entry is zero. To startup the
corresponding processors, the SMP initialization code will update its
entry and let it point to the _start function. To gain access to the
address of the spintable entry, the device tree is parsed.